requirements.txt
Django==3.2.4
gunicorn
djangorestframework
markdown    
django-filter
django-guardian
django-cors-headers
requests
six
djangorestframework-simplejwt
django-password-validators
psycopg2-binary
python-dotenv
Pillow
